Minor client refactoring, css mostly
