+++ client/src/App.js
@@ -1,20 +1,8 @@
 import React from "react";
-
-import "./App.css";
 import Controls from "./components/Controls";
+import SongInfo from "./components/SongInfo";
 import useMpdInfo from "./hooks/mpdInfo";
 import useArtistImage from "./hooks/artistImage";
-
-function SongInfo({ mpdInfo: { artist, album, title } }) {
-  if (!(artist || title)) return null;
-  return (
-    <div id="mpd-info" style={{ textAlign: "left" }}>
-      <h1>{title}</h1>
-      <h2>{artist}</h2>
-      <h3>{album}</h3>
-    </div>
-  );
-}
 
 function App() {
   const { mpdInfo } = useMpdInfo();
@@ -23,8 +11,10 @@ function App() {
 
   return (
     <div className="App" style={{ backgroundImage }}>
-      <SongInfo mpdInfo={mpdInfo} />
-      <Controls currentState={mpdInfo.state} />
+      <div className="wrapper">
+        <SongInfo mpdInfo={mpdInfo} />
+        <Controls currentState={mpdInfo.state} />
+      </div>
     </div>
   );
 }
